+++ packages/zowe-explorer/__mocks__/mockCreators/datasets.ts
@@ -70,6 +70,8 @@ export function createDatasetTree(sessionNode: ZoweDatasetNode, treeView: any, f
         getProfileName: jest.fn(),
         getSession: jest.fn(),
         getProfiles: jest.fn(),
+        getDsTemplates: jest.fn(),
+        addDsTemplate: jest.fn(),
     };
     testDatasetTree.addFavorite.mockImplementation((newFavorite) => testDatasetTree.mFavorites.push(newFavorite));
     testDatasetTree.addFileHistory.mockImplementation((newFile) => testDatasetTree.mFileHistory.push(newFile));